requirements.txt
matplotlib==3.7.1
seaborn==0.13.1
numpy==1.26.4
scipy==1.10.1
statsmodels==0.14.0
pandas==1.5.3
umap-learn==0.5.3
plotly==5.13.1
pytest-cov==4.0.0
pytest-mock==3.10.0
multiprocess==0.70.15
anndata==0.10.0
networkx==3.1
nxontology==0.5.0
netgraph==4.12.11
tqdm==4.65.0
scanpy==1.9.5
ipywidgets==8.0.6

contourpy==1.2.0
cycler==0.12.1
Cython==0.29.36
fonttools==4.45.1
hdbscan==0.8.33
igraph==0.11.3
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.1
numba==0.58.1
packaging==23.2
Pillow==10.0.0
pynndescent==0.5.11
pyparsing==3.1.1
PySimpleGUI==4.60.5
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
six==1.16.0
tenacity==8.2.3
texttable==1.7.0
threadpoolctl==3.2.0
tzdata==2023.3